# make -O1 a bit more useful by disabling more optimization passes