@vtjnash 2018-05-08
For benchmark, might be more interesting to see something that stresses compilation time a bit more, such as `@code_llvm @code_warntype(1 + 1)`(#26714)