@JeffBezanson 2018-05-08
Also, note that since the SLP vectorizer is now enabled by default, there's almost no difference between -O2 and -O3. We might want to remove one of those levels, or else find other passes to move from -O2 to -O3.